Verify user in authorized routes

+++ api/src/routes/memories.ts
@@ -47,7 +47,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         })
     })
 
-    app.get('/memories/:id', async (request) => {
+    app.get('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
@@ -60,10 +60,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
             },
         })
 
+        if (!memory?.isPublic && memory?.userId !== request.user.sub) {
+            return reply.status(401).send()
+        }
+
         return memory
     })
 
-    app.put('/memories/:id', async (request) => {
+    app.put('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
@@ -78,7 +82,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
         const { content, isPublic, coverUrl } = bodySchema.parse(request.body)
 
-        const memory = await prisma.memory.update({
+        let memory = await prisma.memory.findUnique({
+            where: {
+                id,
+            },
+        })
+
+        if (memory?.userId !== request.user.sub) {
+            return reply.status(401).send()
+        }
+
+        memory = await prisma.memory.update({
             where: {
                 id,
             },
@@ -92,12 +106,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
         return memory
     })
 
-    app.delete('/memories/:id', async (request) => {
+    app.delete('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
 
         const { id } = paramsSchema.parse(request.params)
+
+        const memory = await prisma.memory.findUnique({
+            where: {
+                id,
+            },
+        })
+
+        if (memory?.userId !== request.user.sub) {
+            return reply.status(401).send()
+        }
 
         await prisma.memory.delete({
             where: {
